block brakes on dead, increase bomb delay and smooth

+++ ogre/src/cheats/CheatBomb.h
@@ -38,7 +38,7 @@ private:
     bool mIsBombInProgress;
     bool mIsBombJumpsInProgress;
     bool mIsBombExplosionInProgress;
-    Ogre::uint8 mBlowCounter;
+    Ogre::uint16 mBlowCounter;
     Ogre::uint8 mExplosionCounter;
     PhysicsVehicle * mPlayerVehicle;
